Given int value

+++ quantity-store/src/main/java/com/damjan/quantitystore/controller/ItemController.java
@@ -22,7 +22,7 @@ public class ItemController {
     }
 
     @GetMapping("/{sku}")
-    public Item getItemBySku(@PathVariable Long sku) {
+    public Item getItemBySku(@PathVariable Integer sku) {
         return itemService.getItemBySku(sku);
     }
